drf-api/requirements.txt
pip>=24.0.0
Django>=5.1.1
djangorestframework>=3.15.2
sqlparse>=0.5.1
tzdata>=2024.2
